put make packets func in utils
