+++ Phase2/UDPserver.py
@@ -4,6 +4,7 @@ import time
 import sys
 import os
 from collections import deque
+import utils
 
 UDP_IP = "127.0.0.1"	# Server IP
 IN_PORT = 12001	    # port
@@ -14,14 +15,6 @@ if len(sys.argv) < 2:
     sys.exit(1)
 
 response_filename = sys.argv[1]
-
-def make_packets(fileobj, byte_per_pack):
-    data = fileobj.read(byte_per_pack)
-    packets = deque()
-    while data:
-        packets.append(data)
-        data = fileobj.read(byte_per_pack)
-    return packets
 
 # function to send an image
 def send_img(socket, client, filepath):
@@ -35,13 +28,15 @@ def send_img(socket, client, filepath):
     socket.sendto(filename.encode('utf-8'), client)
 
     # read 1024 bytes
-    packets = make_packets(f, 1024)
+    packets = utils.make_packets(f, 1024)
 
     # while theres data keep sending
     print("Sending response file")
     while packets:
         if not socket.sendto(packets.popleft(), client):
             print('Error sending packet')
+    
+    f.close()
             
 sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)		# open a UDP socket
 sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)	# set the socket options to reuse address and port
